+++ src/test/java/com/global/api/tests/portico/PorticoCreditTests.java
@@ -950,5 +950,18 @@ public class PorticoCreditTests {
 
         //TODO: This is insufficient to actually test. The request needs to be checked for the presence of <AmountIndicator>E</AmountIndicator>
     }
+
+    @Test
+    public void credit_Sale_AmountIndicator_Null() throws ApiException {
+        Transaction response = card.charge(new BigDecimal(10))
+                .withCurrency("USD")
+                .withAllowDuplicates(true)
+                .execute();
+        assertNotNull(response);
+        assertEquals("00", response.getResponseCode());
+
+        // TODO: This is insufficient to actually test.
+        // TODO: The request needs to be checked for the absence of the <AmountIndicator/> tag
+    }
 }